+++ operator/api/aistore/v1beta1/aistore_validation.go
@@ -159,7 +159,6 @@ func (ais *AIStore) validateDeprecatedFields() (admission.Warnings, error) {
 		field       string
 		replacement string
 	}{
-		{ais.Spec.ClusterDomain != nil, "spec.clusterDomain", "the operator's configured or discovered cluster domain"},
 		{ais.Spec.HostpathPrefix != nil, "spec.hostpathPrefix", "spec.stateStorage.hostPath.prefix"},
 		{ais.Spec.StateStorageClass != nil, "spec.stateStorageClass", "spec.stateStorage.pvc.storageClass"},
 	}