+++ operator/internal/resources/aistore/cmn/url_test.go
@@ -2,31 +2,37 @@
  * Copyright (c) 2026, NVIDIA CORPORATION. All rights reserved.
  */
 
-package cmn
+package cmn_test
 
 import (
 	"fmt"
 
 	aisapc "github.com/NVIDIA/aistore/api/apc"
 	aisv1 "github.com/ais-operator/api/aistore/v1beta1"
-	"github.com/ais-operator/internal/opinfo"
+	"github.com/ais-operator/internal/resources/aistore/cmn"
+	"github.com/ais-operator/internal/svcaddr"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/intstr"
 )
 
 const (
-	urlTestDomain       = "k8s.example.com"
 	urlTestPublicPort   = 51080
 	urlTestControlPort  = 51082
 	urlTestScheme       = "http"
 	urlTestSecureScheme = "https"
+	urlTestName         = "ais"
+	urlTestNamespace    = "ais-ns"
+	urlTestPrimaryPod   = "ais-proxy-0"
 )
+
+var urlTestProxySvc = types.NamespacedName{Name: "ais-proxy", Namespace: urlTestNamespace}
 
 func newURLTestAIS() *aisv1.AIStore {
 	return &aisv1.AIStore{
-		ObjectMeta: metav1.ObjectMeta{Name: "ais", Namespace: "ais-ns"},
+		ObjectMeta: metav1.ObjectMeta{Name: urlTestName, Namespace: urlTestNamespace},
 		Spec: aisv1.AIStoreSpec{
 			ProxySpec: aisv1.DaemonSpec{
 				ServiceSpec: aisv1.ServiceSpec{
@@ -38,42 +44,26 @@ func newURLTestAIS() *aisv1.AIStore {
 	}
 }
 
-var _ = Describe("ClusterDomain", Label("short"), func() {
-	It("should fall back to the domain of the cluster the operator runs in", func() {
-		Expect(ClusterDomain(newURLTestAIS())).To(Equal(opinfo.ClusterDomain()))
-	})
-
-	It("should prefer the domain from the spec", func() {
-		ais := newURLTestAIS()
-		//nolint:staticcheck // ClusterDomain spec option will be removed in the next major release.
-		ais.Spec.ClusterDomain = aisapc.Ptr(urlTestDomain)
-
-		Expect(ClusterDomain(ais)).To(Equal(urlTestDomain))
-	})
-})
-
 var _ = Describe("Proxy URLs", Label("short"), func() {
 	var ais *aisv1.AIStore
 
 	BeforeEach(func() {
 		ais = newURLTestAIS()
-		//nolint:staticcheck // ClusterDomain spec option will be removed in the next major release.
-		ais.Spec.ClusterDomain = aisapc.Ptr(urlTestDomain)
 	})
 
 	It("should address the primary proxy pod on the intra-control port", func() {
-		Expect(DefaultProxyURL(ais)).To(Equal(fmt.Sprintf("%s://%s.%s.%s.svc.%s:%d",
-			urlTestScheme, ais.DefaultPrimaryName(), ais.ProxyStatefulSetName(), ais.Namespace, urlTestDomain, urlTestControlPort)))
+		Expect(cmn.DefaultProxyURL(ais)).To(Equal(fmt.Sprintf("%s://%s.%s:%d",
+			urlTestScheme, urlTestPrimaryPod, svcaddr.ServiceFQDN(urlTestProxySvc), urlTestControlPort)))
 	})
 
 	It("should address the proxy service on the intra-control port", func() {
-		Expect(DiscoveryProxyURL(ais)).To(Equal(fmt.Sprintf("%s://%s.%s.svc.%s:%d",
-			urlTestScheme, ais.ProxyStatefulSetName(), ais.Namespace, urlTestDomain, urlTestControlPort)))
+		Expect(cmn.DiscoveryProxyURL(ais)).To(Equal(fmt.Sprintf("%s://%s:%d",
+			urlTestScheme, svcaddr.ServiceFQDN(urlTestProxySvc), urlTestControlPort)))
 	})
 
 	It("should address the proxy service on the public port", func() {
-		Expect(IntraClusterURL(ais)).To(Equal(fmt.Sprintf("%s://%s.%s.svc.%s:%d",
-			urlTestScheme, ais.ProxyStatefulSetName(), ais.Namespace, urlTestDomain, urlTestPublicPort)))
+		Expect(cmn.IntraClusterURL(ais)).To(Equal(fmt.Sprintf("%s://%s:%d",
+			urlTestScheme, svcaddr.ServiceFQDN(urlTestProxySvc), urlTestPublicPort)))
 	})
 
 	It("should use https when the cluster does", func() {
@@ -83,6 +73,6 @@ var _ = Describe("Proxy URLs", Label("short"), func() {
 			},
 		}
 
-		Expect(IntraClusterURL(ais)).To(HavePrefix(urlTestSecureScheme + "://"))
+		Expect(cmn.IntraClusterURL(ais)).To(HavePrefix(urlTestSecureScheme + "://"))
 	})
 })