+++ operator/internal/resources/aistore/proxy/statefulset.go
@@ -61,7 +61,7 @@ func NewProxyStatefulSet(ais *aisv1.AIStore, size int32) *apiv1.StatefulSet {
 			Selector: &metav1.LabelSelector{
 				MatchLabels: SelectorLabels(ais),
 			},
-			ServiceName:          headlessSVCName(ais.Name),
+			ServiceName:          ais.ProxyHeadlessSVCNSName().Name,
 			PodManagementPolicy:  apiv1.ParallelPodManagement,
 			Replicas:             &size,
 			VolumeClaimTemplates: proxyVC(ais),
@@ -134,7 +134,7 @@ func proxyPodSpec(ais *aisv1.AIStore) *corev1.PodSpec {
 
 func NewInitContainerEnv(ais *aisv1.AIStore) (initEnv []corev1.EnvVar) {
 	initEnv = cmn.CommonInitEnv(ais, ais.ProxyExternalAccessEnabled())
-	initEnv = append(initEnv, cmn.EnvFromValue(cmn.EnvServiceName, headlessSVCName(ais.Name)))
+	initEnv = append(initEnv, cmn.EnvFromValue(cmn.EnvServiceName, ais.ProxyHeadlessSVCNSName().Name))
 	// Set AIS_PUBLIC_HOSTNAME if listening on hostPort
 	// Without this set, the proxy will resolve its public hostname to the pod IP
 	if ais.Spec.ProxySpec.HostPort != nil {