+++ operator/internal/services/client_manager.go
@@ -17,6 +17,7 @@ import (
 
 	aisv1 "github.com/ais-operator/api/aistore/v1beta1"
 	aisclient "github.com/ais-operator/internal/client"
+	"github.com/ais-operator/internal/resources/aistore/cmn"
 	"github.com/ais-operator/internal/resources/aistore/proxy"
 	"github.com/go-logr/logr"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
@@ -135,27 +136,21 @@ func logNewClient(logger logr.Logger, tokenInfo *TokenInfo, tlsConf *tls.Config,
 func (m *AISClientManager) getAISAPIEndpoint(ctx context.Context,
 	ais *aisv1.AIStore,
 ) (string, error) {
-	var (
-		err      error
-		hostname string
-		port     string
-	)
-
 	switch {
 	case ais.GetAPIMode() == APIModePublic:
-		hostname, err = m.getPublicAISHostname(ctx, ais)
+		hostname, err := m.getPublicAISHostname(ctx, ais)
 		if err != nil {
 			logf.FromContext(ctx).Error(err, "Failed to get public AIS API parameters")
 			return "", err
 		}
-		port = ais.Spec.ProxySpec.PublicPort.String()
+		return cmn.AISHostURL(ais, hostname, ais.Spec.ProxySpec.PublicPort.String()), nil
 	// If LoadBalancer is configured use the LB service to contact the API.
 	case ais.ProxyExternalAccessEnabled():
 		proxyLBSVC, svcErr := m.k8sClient.GetService(ctx, proxy.LoadBalancerSVCNSName(ais))
 		if svcErr != nil {
 			return "", svcErr
 		}
-
+		var hostname string
 		for _, ing := range proxyLBSVC.Status.LoadBalancer.Ingress {
 			if ing.IP != "" {
 				hostname = ing.IP
@@ -169,14 +164,12 @@ func (m *AISClientManager) getAISAPIEndpoint(ctx context.Context,
 		if hostname == "" {
 			return "", fmt.Errorf("proxy load balancer svc %q has no ingress IP or hostname", proxy.LoadBalancerSVCNSName(ais))
 		}
-		port = ais.Spec.ProxySpec.ServicePort.String()
+		return cmn.AISHostURL(ais, hostname, ais.Spec.ProxySpec.ServicePort.String()), nil
 	// When operator is deployed within K8s cluster with no external LoadBalancer,
 	// use the proxy headless service to request the API.
 	default:
-		hostname = proxy.HeadlessSVCNSName(ais).Name + "." + ais.Namespace
-		port = ais.Spec.ProxySpec.ServicePort.String()
+		return cmn.IntraClusterURL(ais), nil
 	}
-	return createAPIURL(ais.UseHTTPS(), hostname, port), nil
 }
 
 func (m *AISClientManager) getPublicAISHostname(ctx context.Context, ais *aisv1.AIStore) (hostname string, err error) {
@@ -318,12 +311,4 @@ func findCerts(root string, exts []string) ([]string, error) {
 		return nil
 	})
 	return certPaths, err
-}
-
-func createAPIURL(https bool, hostname, port string) string {
-	scheme := "http"
-	if https {
-		scheme = "https"
-	}
-	return fmt.Sprintf("%s://%s:%s", scheme, hostname, port)
 }